% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{imperial_pal}
\alias{imperial_pal}
\title{Return function to interpolate an Imperial color palette}
\usage{
imperial_pal(palette = "core", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in imperial_palettes. Choose from:
\itemize{
 \item core, cool, warm, extended,
 \item pink, green, red blue,
 \item teal_blue_plum, violet_blue_navy, orange_green_imperial_blue, brick_navy_teal,
 \item blue_grey_gradient, blue_green_gradient
}}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate an Imperial color palette
}
