% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{scale_color_imperial}
\alias{scale_color_imperial}
\title{Color scale constructor for imperial colors}
\usage{
scale_color_imperial(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in imperial_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to ggplot2::discrete_scale() or
ggplot2::scale_color_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Color scale constructor for imperial colors
}
