% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{imperial_pal}
\alias{imperial_pal}
\title{Return function to interpolate an Imperial color palette}
\usage{
imperial_pal(palette = "core", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in imperial_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to dichromat::colorRampPalette()}
}
\description{
Return function to interpolate an Imperial color palette
}
