% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{scale_fill_imperial}
\alias{scale_fill_imperial}
\title{Fill scale constructor for imperial colors}
\usage{
scale_fill_imperial(palette = "core", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in imperial_palettes. Choose from:
\itemize{
 \item core, cool, warm, extended,
 \item pink, green, red blue,
 \item teal_blue_plum, violet_blue_navy, orange_green_imperial_blue, brick_navy_teal,
 \item blue_grey_gradient, blue_green_gradient
}}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to ggplot2::discrete_scale() or
ggplot2::scale_fill_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Fill scale constructor for imperial colors
}
